requirements.txt
paho_mqtt==2.1.0
Requests==2.32.3
streamlit==1.39.0
tqdm==4.66.5
